+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+0.1.7-alpha:
+ - Home text edit
+ - Tooltips in sync section
+ - Removed unused fbuser var from part.settings
+
 0.1.6-alpha:
- + Home text
- ~ Fixed crop for picture smaller than 400px
+ - Home text edit
+ - Fixed crop for picture smaller than 400px
 
 0.1.5-alpha:
- ~ First release
+ - First release